Require explicitly listing xtest files separately
There's no way to know if a test is internal or external without opening
it. Instead of always trying to compile both and adding some special
case to filter source files at compile time, make it explicit. It's a
little more typing but allows skipping work. With dynamic derivations,
this could eventually be automated.